v1/requirements.txt
msal>=1.28
python-dotenv>=1.0
requests>=2.31
pdfplumber>=0.11
pymupdf>=1.24
pytesseract>=0.3.10
Pillow>=10.0
pydantic>=2.0
pandas>=2.0
PyYAML>=6.0
openpyxl>=3.1
xlrd>=2.0
anthropic>=0.40
tenacity>=8.2
